Hand over to community earlier if high risk case
I missed in 39134a that the behaviour for MAPPA handovers is unusual
- both handover start and handover itself happen at 7.5 months before
the sentence ends at the latest.

With this implementation, the actual handover date will change each
day after a MAPPA flag has been set (if the offender's release date is
less than 7.5 months away), but the handover start date will still be
correct. From a steer from our PO, this is fine.