Add a new summary page for handover cases
This commit adds a new 'Handover cases' summary page. This page will be used to display offenders who are approaching their responsibility handover date.

The page content itself is currently a work in progress. This commit simply adds the new page into the application.

I've chosen to extend the existing summary pages, which means using the summary controller and adapting the summary service and presenter objects. It makes most sense for the new page to belong to this collection of existing pages, since it is mostly a new view on existing data which is also used on these other pages.

- Add a route for the new page (`/summary/handovers`) to `config/routes.rb`
- Add a new `handovers` method to the summary controller
- Update the summary model, service and presenter to support the new 'handover' summary type
- Add a view for the new page
- Add the new page to the summary subnav partial