Add logic to find offenders approaching their handover date
I've added a new method `approaching_handover?` to the `SummaryService`. This is used to determine whether an offender is approaching their handover date, or is already within the handover window (i.e. the handover period has started, as determined by the 'handover start date', but has yet to finish, as determined by the 'responsibility handover date'). This may not be the best place for such a method, as it could also be useful on the `Offender` model itself – however, for now I've added it here to be consistent with the existing `new_arrival?` filter.

The new `approaching_handover?` method is used within the offender loop (in the `self.summary` method) to put offenders that are approaching their handover date into the `:handovers` bucket. This is consistent with the existing approach for other summary views on offenders (i.e. allocated, unallocated, pending more information, new arrivals).

I'd like to test this method somehow to ensure it filters offenders correctly. However I'm committing this as-is for now, until we've been able to spend some time thinking about our testing strategy for this service object.