Change columns on the 'Handover cases' page
Following a conversation within the team, we've decided to remove the 'Earliest release date' column from this page as it is not entirely useful in this view. Instead, we've added a column to show the assigned COM for an offender.

The assigned COM name is not (currently) available in the dataset provided to this view, so for now I've put a 'TODO' placeholder in this column.