Show allocated COM name on 'Handover cases' page
If allocated, the COM name is shown on the 'Handover cases' page for each offender.

To achieve this, I've had to add a new custom attribute to both the `OffenderBase` model and the `OffenderPresenter` presenter objects. I'm not entirely sure how I feel about this as a global change, since this field is only needed on one view within the application – it certainly feels like there could be a better way to do this.