Set the default sort order for 'Handover cases'
The 'Handover cases' page is now sorted by 'Handover start date' by default. This differs from other summary pages, which tends to default to sort by prisoner name.

In the process of this, I refactored the private method `sort_params` in the summary controller in an attempt to make it clearer. In doing so, I introduced a new method `default_sort_params` which is solely responsible for the default sort order of summary pages.

The sort arrow in the UI always defaults to the 'Prisoner' (`:last_name`) column, because the sort helper has its own defaults independent of how it's handled in the controller. I'm not sure how best to rationalise this.