Refactor `Summary` class constructor
Rubocop was showing an offence due to the number of parameters required by the `Summary` class constructor.

To simplify the constructor parameters, I've refactored it to accept a Hash of offender buckets created by the `SummaryService`, rather than those buckets being passed through individually.

Since `SummaryService` is the only class that creates `Summary` objects, it was quite a simple refactor.