Add link to allocation info page from handovers
I've added a link to view allocation information from the 'Handover cases' page.

In most cases, this will display as a 'View' link. However if the offender has not been allocated, it will display an 'Allocate' link. This is consistent with the behaviour on the search results page.

In order to squeeze in an extra column on this page, Eve and I decided to drop the full-length formatting of the service provider name. So where this column used to read "National Probation Service (NPS)", for example, it will now simply display "NPS". This reduces the cell content enough to be able to fit the new column in.