Always enable VCR in tests (no effect on webmock)
This commit changes our rspec configuration, so that VCR is always
enabled - meaning our build server's test run now matches local test
runs.

I hope this'll tighten the feedback loop with testing, and avoid "I
thought this was fine but it's broken on the build server" pain.

I dug through commit history a little to try and determine the reason
for the distinct behaviour originally - I think it's the remnants of
experimenting with webmock/VCR, but it's possible I've missed
something. If so, it's cheap to revert this change.

Note that this doesn't change the VCR/webmock interactions - this
just makes it so VCR is always in context, but it may still defer to
webmock on a test-by-test basis. At the time of writing, I believe
this should happen when no casette is specified, but have seen other
behaviour that undermines that understanding.