DRY-up loop code in the Delius import job
I've DRY-ed up the code in the Delius import job which:
- loops through every row of the Delius spreadsheet
- skipping the first (header) row
- and maps the row Array into a Hash with keys for field names

This loop is being performed twice, as we now perform two passes of the
Delius extract spreadsheet – once to update team names & LDUs, and once
more to import offender data.

I've extracted the record looping functionality into a method called
`for_each_record`, which yields the block passed into it. That way we
can control what action to perform on each loop.