Refactor method to reduce code complexity
I've refactored the `UpdateTeamNameAndLduService#update` method to
resolve a code complexity warning in CodeClimate.

I needed to refactor one of the unit tests as I realised it was testing
for the wrong behaviour. I was asserting that the team record is not
saved if the associated LDU is not found. However, this should only be
true if the record hasn't changed for any other reason – i.e. if the
team name was already correct.

I've also taken the opportunity to clarify the wording of a couple of
context block descriptions which test this and related behaviour.