Remove duplicate certificate deployment config
This commit removes the `certificate.yaml` deployment configuration file
from our codebase. It is not required in this repository as it also
exists in the Cloud Platform environments repository:

https://github.com/ministryofjustice/cloud-platform-environments/blob/master/namespaces/live-1.cloud-platform.service.justice.gov.uk/offender-management-production/certificate.yaml

[The commit which created that configuration file](https://github.com/ministryofjustice/cloud-platform-environments/commit/db499c8ee2ee3831dc5b3eea7a060a9ba39e89e2)
 suggests we'd intended to remove the equivalent file from this
repository – I'm not entirely sure why that didn't happen.

Rather than maintain our own copy of that configuration file in this
repo, it seems more sensible for us to rely on the one
which exists in the Cloud Platform repository:

Firstly, it removes the ambiguity of having duplicate configuration
files in both places – there will only be one 'source of truth' for
certificate configuration.

Secondly, the Cloud Platform environments repository is more likely to
be supported by the Cloud Platforms team. For example, with the recent
changes to how certificates are deployed, the Cloud Platform team
[automatically updated](https://github.com/ministryofjustice/cloud-platform-environments/commit/a444add05384f8482c80d23aa1b0953a09add338#diff-fb1e255aece5dd0ea5e56073c29c006a)
 the `certificate.yaml` file for us. However, because we didn't copy
those changes across to our own repo, they were being overridden by the
out-of-date config file in this repository – ultimately leading to our
deployments breaking.

By removing this file from our repository, deployments will fall back to
 the configuration file which is maintained in the Cloud Platform
 environments repository.