Update shadow codes for teams
This commit builds upon the Delius import job to allow our service to
automatically 'learn' about shadow team codes.

Our modelling of active and shadow teams is slightly different from the
way they're modelled in Delius.

In Delius, two types of team exist:
- Active teams
- Shadow teams

They're differentiated by the LDU they belong to – if a team's LDU code
looks like `N01OMIC` (where `01` is any 2 digit number), then it's a
shadow team. Otherwise, it's an active team.

In our application, however, there's only one type of team. Instead,
teams have both an 'active' code (field: `code`), and a 'shadow' code
(field: `shadow_code`).

Delius has no concept of relationships between teams, however there's a
convention which dictates that shadow teams share the same name as their
active counterpart, but with 'OMIC' prepended.

Knowledge of the relationship between 'active' and 'shadow' teams is
something that this service is responsible for.

This commit introduces a new service object called
`UpdateShadowTeamAssociation`. This new service receives a Delius
'shadow' team name and code, and attempts to locate the associated
'active' team in our database. The lookup is performed based on the
aforementioned shadow team naming convention, which should allow us to
locate the active team.

If it finds a team with the target name, it'll update that team's shadow
code to reflect the one seen in the Delius import.

In essence, this will allow us to automatically learn about shadow team
codes provided we already have the active team in our database.