Update teams from Delius according to their type
When iterating over the Delius extract spreadsheet in the nightly import
job, we want to update our own knowledge of the relationships between
shadow teams, active teams and LDUs.

For each row in the Delius spreadsheet, a different action will be
performed depending on the type of team in that row:

Active teams are identified as having an LDU code which *does not* look
like `N01OMIC`.

For these teams, Delius is the source of truth for which LDU the team
belongs to, as well as the name of the team.

Therefore, when we see an active team, we'll process it using the
`UpdateTeamNameAndLduService` service.

This will update our knowledge of which active teams belongs to which
LDUs.

Shadow teams are identified as having an LDU code which looks like
`N01OMIC` (where `01` could be any 2 digit number).

For these teams, Delius is *not* the source of truth for which LDU the
team belongs to. Instead, we perform our own mapping of
shadow team => active team, which is what ultimately determines the LDU
for that shadow team.

Therefore, when we see a shadow team, we'll process it using the
`UpdateShadowTeamAssociationService`.

This will update our knowledge of which shadow codes belong to which
teams.

---

We perform these updates in two passes over the Delius spreadsheet:

1) Process active teams
2) Process shadow teams

When done in this order, we'll have the best chance of capturing data
for both active and shadow teams.