Add Team has many Case Information association
The Case Information model already had a 'belongs to' association to the
Team model. However, the inverse of that association hadn't been
established – a Team 'has many' Case Information records. This commit
adds that association.

I'm adding this now because I think it'll be useful when adding an
offender count to the teams section of the active admin interface.