Make teams sortable by the number of offenders
The "Offenders" column on the Active Admin "Teams" page is now sortable.

To do this, I've added a counter cache to the association between Teams
and Case Information records. By using a counter cache for the number of
offenders (case information) records assigned to each team, we can
easily perform sorts against that field in Active Admin.

This will be useful as we try to identify unused teams which could be
removed from our database.

The database migration which creates the new `case_information_count`
column will also auto-populate the counts for existing team records –
this happens as part of the migration.