Add index on teams.shadow_code column
We regularly find teams by their shadow code as part of our application.
In particular, we do this as part of our nightly Delius import job.

We already have an index on the `code` column in the `teams` database
table, so it makes sense to also add an index on the `shadow_code`
column to support this common read pattern.