Fix lint error in rake task
Our Rubocop lint rules enforce that all rake tasks should require the
`:environment` task. This was triggering the cop
[Rails/RakeEnvironment](https://docs.rubocop.org/rubocop-rails/cops_rails.html#railsrakeenvironment).

The line itself which triggered this error seemed somewhat confusing to
me. After some searching around on the web, I found that this
confusing-looking line is actually a workaround for rake's sub-optimal
handling of rake task arguments.

In understanding the different options for accepting arguments in rake
tasks, I found this article really helpful: https://cobwwweb.com/4-ways-to-pass-arguments-to-a-rake-task

In it, the author discusses the meaning of that `ARGV.each` line. In
short, it dynamically creates 'dummy' tasks which do nothing, but are
named after the values passed in `ARGV`. This is because rake will
otherwise assume the second parameter passed is another task to run,
and will throw an error without the dummy task existing.

In this case, the fix was to ensure that these dummy tasks still require
the `:environment` task, to satisfy the `Rails/RakeEnvironment` cop.