Enable CRC offenders to display correctly
This commit creates 2 tests, one that checks that the CRC and NPS Cases
are within the 30 day window before their release for the
summaryController and one for the caseLoadController. If the offender
is in the window, they are returned, allowing the them to be included
on the POM 'Cases close to handover' page and the SPO 'Case handover
status' page. In order for the tests to pass, the handover method,
inside HandoverDateService, was modified so that CRC’s start_date,
which had been set to nil, was replaced with crc_handover_date - t
his date now acting as a start date (in order to be displayed,
offenders are required to have a start date).

Much of the test data was modified in order order to create a CRC
offender that sits within the 30 day window.