Add rake task for seeing missing handover emails (#1153)
* Add rake task for seeing missing handover emails

This commit includes a new class (what I'd call a 'use case', but is
commonly a 'service' in this app), tests and a custom rake task that
can be run with `rake handover_email_summary_by_prison [prison_code]`.

Notably, this rake task cannot give information across the entire
estate, only for an individual prison. I expect it's possible to get
a view across the state in one task with some fault-tolerance /
batch-processing work.

These counts are not battle-tested yet. I expect there is at least one
problem - I haven't yet seen a way to differentiate between:
- This record is not marked as 'NPS' because it doesn't have the
relevant Delius record stored against it yet.
- This record is genuinely not marked as 'NPS'.

* Use large elseif instead of multiple returns

These weren't really guard clauses, so multiple returns could be less
expressive (and notably, more likely to confuse if the branches grow
over time).

With the various conditions split off into other methods, I think the
size of this elsif is justifiable, and expresses the idea of 'try
this bucket, then the next one, then the next one' pretty well.

This pattern has some repetition, but I think metaprogramming and
using the method names won't pay off readability-wise.

* Fix lint error in rake task

Our Rubocop lint rules enforce that all rake tasks should require the
`:environment` task. This was triggering the cop
[Rails/RakeEnvironment](https://docs.rubocop.org/rubocop-rails/cops_rails.html#railsrakeenvironment).

The line itself which triggered this error seemed somewhat confusing to
me. After some searching around on the web, I found that this
confusing-looking line is actually a workaround for rake's sub-optimal
handling of rake task arguments.

In understanding the different options for accepting arguments in rake
tasks, I found this article really helpful: https://cobwwweb.com/4-ways-to-pass-arguments-to-a-rake-task

In it, the author discusses the meaning of that `ARGV.each` line. In
short, it dynamically creates 'dummy' tasks which do nothing, but are
named after the values passed in `ARGV`. This is because rake will
otherwise assume the second parameter passed is another task to run,
and will throw an error without the dummy task existing.

In this case, the fix was to ensure that these dummy tasks still require
the `:environment` task, to satisfy the `Rails/RakeEnvironment` cop.

* Fix lint errors in spec file

This commit fixes Rubocop errors flagged up by cops
`RSpec/DescribeClass` and `RSpec/ContextWording` in the spec file for
the 'view handover email addresses summary' service object.

* Fix naming inconsistency: 'delivery'/'divisional'

This commit resolves a naming inconsistency in the 'view handover email
addresses summary' service, which referred to LDUs as 'Local Delivery
Units', whereas elsewhere in the codebase they're known as 'Local
Divisional Units'.

I've seen mixed messages throughout HMPPS as to what, exactly, the LDU
acronym should be expanded to. I have a suspicion that our use of
'Local Divisional Unit' is actually incorrect. However I've made this
change for the sake of consistency throughout our codebase.

* Add comment to explain purpose of service

I've added a comment to the `ViewHandoverEmailAddressSummary` service to
explain what it does and how to use it.

Co-authored-by: Ollie Treend <ollie@treend.com>
Co-authored-by: Ollie Treend <ollie.treend@digital.justice.gov.uk>