Remove old XLSX file reader
This commit removes `Delius::Processor` and associated components.

This was a custom implementation of a XLSX file reader, but it had a bug
where it would incorrectly read the values of cells containing an
ampersand. We now use the gem `roo` to read XLSX spreadsheets.

These components were used to read the Delius extract file (which is a
XLSX Excel spreadsheet) and import data into our service. All uses of
these components have now been swapped out for
`Delius::ExtractFileReader`, meaning it's now safe to remove them from
our codebase.