Pad empty cells when importing the Delius extract
Without explicitly enabling the `pad_cells` option of `roo` when reading
the Delius XLSX file, it will skip empty cells as if they don't exist.

In practice, there are sometimes empty cells in the Delius XLSX files
that we import. Without padding those cells, there's a risk they'll
be mapped to the wrong field names and break things in the import.

For example, when importing a spreadsheet row which has an empty "C"
column, you might end up with an array which looks something like:

```
["A", "B", "D", "E"]
```

Note that the "C" column is missing, but it's also not being represented
as being empty. It's simply skipped altogether.

By padding empty cells, that same example would result in an array like:

```
["A", "B", nil, "D", "E"]
```

The important thing here is that the index of cell values remains
intact. Without that, we'd fail at mapping cell values to their
corresponding field names – as we'd have no way to know that the "C"
column is empty.

In `Delius::ExtractFileReader`, empty cells are represented as empty
strings rather than as `nil` values. So when we map over the row to
produce an array of cell values, we rewrite `nil` (which we receive from
roo) as `""`. This seems to be the most representation of empty cells
for our needs because it looks like they somehow already import as empty
strings in real-world export files – so best to codify that behaviour.