Ignore blank rows in the Delius import
The last row of the Delius extract spreadsheet tends to be completely
blank – i.e. every cell is empty.

We should ignore blank rows in the spreadsheet, since they won't contain
the data we're expecting and it'll end up hurting us in unpredictable
ways.

This seems to have tripped us up in the past, as the old
`Delius::Processor` implementation wasn't skipping blank rows either.

The test fixture I've added to cover this scenario is possibly a bit
over-the-top in that it contains several blank rows interspersed amongst
'proper' rows of data. In practice, we only tend to see one blank row in
the Delius extract spreadsheet, and it's always the final row.

However, blank rows at the end of the spreadsheet are easy to miss,
which is why I chose to make the test fixture very explicit with how it
demonstrates empty rows.