Skip blank rows in the nightly Delius import
This commit patches a bug in the nightly Delius import where it was crashing when encountering blank rows in the XLSX spreadsheet. It turns out the Delius spreadsheet always tends to end with a blank row, so this was causing the import to fail in production.

This is a quick fix to patch the problem for the current implementation of the Delius XLSX file reader. We are aware of other bugs in the current implementation (see Jira ticket POM-739), so this is a temporary fix until that work is completed.