It displays case type badges:
 determinate, indeterminate, recall and parole eligilbility case_type_badges on relevant pages so that POMs and SPOs can know at a glance what case type the offender is.

In addition, when a parole review date is manually updated on the the 'Allocate a Prison Offender Manager' page then a 'Parole eligible' badge should display.

This commit also increases the margin below the case-type badge on the allocation page so that it looks less squashed.

Factories have been amended so that parole eligibility dates, tariff dates and parole review date now need to be explicitly set in each test so that test data is accurate.