Tighten up CaseInformation model validation rules
I've refactored and tightened up the CaseInformation validation rules.

The primary driver for this was to clarify which field values are
allowed in which cases.

When probation service is England / Wales:
- tier must be A, B, C or D
- case allocation must be NPS or CRC
- a team must be given

When probation service is Scotland / Northern Ireland:
- tier must be N/A (not nil)
- case allocation must be N/A (not nil)
- team will be nil

I've updated the CaseInformation model tests to cover these scenarios.

To help with this, I've conceptually grouped the fields 'tier', 'case
allocation' and 'team' and given them the label of 'probation data' –
i.e. when the probation service is England / Wales, we need this data
about the case, which comes from Delius (the database used by probation)

I've introduced a new method which can be used to determine if those
probation data fields are required: `requires_probation_data?`

By refactoring the CaseInformation model and controller to make use of
this new method, I've been able to DRY-up repeated hardcoded references
to 'England/Wales' and 'Scotland/Northern Ireland' which was being used
to determine user flow, behaviour and validation rules. Hopefully this
helps the readability of the code.