Speed up probation_service database backfill task
The `backfill:probation_service` rake task sets the value of
probation_service field based on the value of welsh_offender column.

This was done by looping through each record individually and performing
a write to the database for each record. This isn't efficient. Given the
number of records we need to backfill, this command would take a long
time to complete.

Instead, I've updated the task to use 'update all'. This performs a SQL
UPDATE statement, which will update all records at once – no need for us
to read/write each record individually.