Only connect to Prometheus Exporter when needed
We set environment variable `PROMETHEUS_METRICS=on` when we want to log
metrics using Prometheus Exporter. However, even when that variable
wasn't set, Puma (web server) and Sidekiq workers (job queue) were still
attempting to initialise a connection to a local Prometheus Exporter
server.

We don't record metrics to Prometheus in Heroku, and therefore don't
run a local Promethus Exporter server. But since the application was
still trying to connect to the server regardless of the environment
variable not being set, it was resulting in the following error being
logged repeatedly in the app logs, at a rate of about two entries per
second:

```
Prometheus Exporter, failed to send message Connection refused - connect(2) for "localhost" port 9394
```

This commit adjusts the Puma and Sidekiq configurations so they don't
attempt to connect to Prometheus Exporter when `PROMETHEUS_METRICS=on`
is not set in the environment. This significantly reduces log noise in
Heroku.