Add API stub for individual POM
I've added a method `stub_pom` to complement the existing method
`stub_poms` in the API spec helper.

The stubs the `/staff/{staffId}` Prison API endpoint which is used to
display the name of the currently allocated POM on the
'Allocation information' page, amongst others.

It accepts the same `:pom` factory object as `stub_poms`, for the sake
of simplicity, even though the endpoints do seem to return slightly
different sets of fields. To make these compatible, I had to add a new
'status' field which is returned by the `/staff/{staffId}` endpoint.