Add stub and factory for the Keyworker API
The method `stub_keyworker` can be used to stub out the Keyworker API.
This API is used on the 'Allocation information' page, amongst others,
to display the key worker assigned to the offender.

The factory `:keyworker` will create an object representing the JSON
response expected from the API, which should be passed in to the
`stub_keyworker` method.

Example usage:

```
stub_keyworker('LEI', 'A01234', build(:keyworker))
```