Add script to copy production database to preprod
This commit introduces a new script which will copy the Postgres database from production into the pre-production environment.

It automates the process which is currently performed manually, and so the hope is that this makes it a less 'scary' operation to perform by reducing chance of human error.

It is based on the steps documented in the Google Doc: [Copy prod db to preprod env][1]

Run the script from your terminal:

```
bin/copy-prod-db-to-preprod.sh
```

You will be asked to confirm before the script begins moving data.

The following commands must be available from the command line:

- `kubectl` – installed and [authenticated to the MOJ Cloud Platform][2]
- [`yq`][3] – needed to parse YAML. Install using homebrew: `brew install yq`

I've tried to add comments to the script to explain what's happening throughout, but here's a high level overview:

1. Using Kubernetes Secrets (available via `kubectl`), get the Postgres RDS database credentials for both the production and pre-production environments.
2. Spin up a pod (container) which has the Postgres CLI, passing in database credentials as environment variables.
3. Execute a bash script on the newly spun-up pod which will copy the database using credentials from the environment.

**Note:** the database copy itself (step 3) runs within the context of a pod in the MOJ Cloud Platform

[1]: https://docs.google.com/document/d/15rP3GFcHFivTvYPlrWqgjwW_R5rM2Emw4e-lYhaAB1U/edit
[2]: https://user-guide.cloud-platform.service.justice.gov.uk/documentation/getting-started/kubectl-config.html
[3]: https://github.com/mikefarah/yq