Restructure HMPPS API class namespaces
This commit restructures the namespacing for classes and modules related
to the HMPPS APIs.

These previously existed in the namespace `Nomis::Elite2`, which no
longer truthfully reflects the names and hierarchy of the APIs:

- The Elite2 API has been renamed to Prison API
- Keyworker API is a separate API, which is not (and never was) part of
the Elite2 / Prison API
- `Nomis::Elite2::Client` was a misleading class name given that it is
intended to be used as a client across all HMPPS APIs.

This new structure should more truthfully reflect the current state of
the world:

- HMPPS API (contains everything related to speaking to a HMPPS API)
  - Client (API client which can be used across any HMPPS API)
  - Prison (grouping of classes which speak to the Prison API)
  - Keyworker (grouping of classes which speak to the Keyworker API)
  - Oauth (grouping of classes concerned with authenticating to these
  APIs)