Add missing stub for Keyworker API client spec
The spec for `HmppsApi::Keyworker::Keyworker` was failing because:
- the spec *does not* use VCR, and
- the auth token had not been stubbed

This meant that tests of this API were attempting to trigger real-world
HTTP queries to the oAuth server to retrieve a valid API access token.

For some reason, this was not failing when run as part of the entire
test suite. There must be a leaky test somewhere which was causing the
auth token to already be stubbed. However it was failing when run in
isolation with:

```
bundle exec rspec spec/services/hmpps_api/keyworker
```