Cache responses properly in the Movement API
This commit alters the way the Movement API is cached.

Previously, hydrated `HmppsApi::Movement` data objects were stored in
the Redis cache. Whilst this mostly worked, it's inconsistent with the
way other API responses are cached, and is therefore considered a bug.

Now, only the data returned from the Movement API is cached. The
subsequent process of 'deserializing' the response into hydrated
`HmppsApi::Movement` objects is always done on-the-fly, uncached. This
is consistent with other API endpoints.

It resolves a recent issue where Movement objects fail to rehydrate from
cache when deploying the recent PR #1333. Because the class name had
changed from `Nomis::Movement` to `HmppsApi::Movement`, the app was
crashing when retrieving objects from the cache.

**Note:** The cache key used by this method has been updated. This is
important due to the fact that we want to invalidate the existing cache
when this is deployed.