Use constants for date codes to help readability
The Community API requires that we specify 'type codes' for handover
dates when we push them into the 'key dates' for an offender.

This mapping was maintained directly in the code, but I've pulled them
up into constants within the `CommunityApi` class. This helps code
readability by allowing us to refer to the date type codes by names that
are more meaningful to us.